+++ packages/toggle-button/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/ToggleButton';
+export * from './lib/ToggleButtonGroup';