+++ packages/tooltip/src/lib/Tooltip.stories.tsx
@@ -71,7 +71,7 @@ export default {
   tags: ['autodocs'],
   argTypes,
   args: {
-    children: 'span',
+    children: 'button',
     title: 'This is a tooltip',
   },
 };