+++ packages/avatar/CHANGELOG.md
@@ -2,13 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-avatar@0.1.3...@availity/mui-avatar@0.1.4) (2023-08-23)
+
+## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-avatar@0.1.2...@availity/mui-avatar@0.1.3) (2023-08-16)
+
+### Bug Fixes
+
+- **mui-avatar:** add alt text to the ImageAvater story for 508 compliance ([29c024c](https://github.com/Availity/element/commit/29c024c80b45e0951a748464ca695aee6ca92e0c))
+
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-avatar@0.1.1...@availity/mui-avatar@0.1.2) (2023-07-13)
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-avatar@0.1.0...@availity/mui-avatar@0.1.1) (2023-06-26)
 
 ## 0.1.0 (2023-06-20)
 
-
 ### Features
 
-* **mui-avatar:** initial commit ([ed356db](https://github.com/Availity/element/commit/ed356dbaeef699b1a8fb201ec3be8c0bc84d62c9))
+- **mui-avatar:** initial commit ([ed356db](https://github.com/Availity/element/commit/ed356dbaeef699b1a8fb201ec3be8c0bc84d62c9))